Move BuildType to constants module
