+++ build-executor/src/main/java/org/jboss/pnc/executor/BuildTypeToRepositoryType.java
@@ -17,7 +17,7 @@
  */
 package org.jboss.pnc.executor;
 
-import org.jboss.pnc.model.BuildType;
+import org.jboss.pnc.enums.BuildType;
 import org.jboss.pnc.model.TargetRepository;
 
 /**
