+++ rest/src/test/java/org/jboss/pnc/rest/endpoint/BuildRecordEndpointTest.java
@@ -21,7 +21,7 @@ import org.jboss.pnc.coordinator.maintenance.TemporaryBuildsCleanerAsyncInvoker;
 import org.jboss.pnc.executor.DefaultBuildExecutionConfiguration;
 import org.jboss.pnc.executor.DefaultBuildExecutionSession;
 import org.jboss.pnc.model.BuildRecord;
-import org.jboss.pnc.model.BuildType;
+import org.jboss.pnc.enums.BuildType;
 import org.jboss.pnc.model.SystemImageType;
 import org.jboss.pnc.model.User;
 import org.jboss.pnc.rest.notifications.websockets.DefaultNotifier;
