+++ rest/src/test/java/org/jboss/pnc/rest/serialization/BuildExecutionConfigurationTest.java
@@ -20,7 +20,7 @@ package org.jboss.pnc.rest.serialization;
 
 import org.jboss.pnc.common.json.JsonOutputConverterMapper;
 import org.jboss.pnc.executor.DefaultBuildExecutionConfiguration;
-import org.jboss.pnc.model.BuildType;
+import org.jboss.pnc.enums.BuildType;
 import org.jboss.pnc.model.SystemImageType;
 import org.jboss.pnc.rest.restmodel.BuildExecutionConfigurationRest;
 import org.jboss.pnc.spi.builddriver.exception.BuildDriverException;
